Only paramaterize the database fixture
